created a resolver and schema files
